+++ SodaMachine/SodaMachine.cs
@@ -112,8 +112,8 @@ namespace SodaMachine
         //If the payment does not meet the cost of the soda: dispense payment back to the customer.
         private void CalculateTransaction(List<Coin> payment, Can chosenSoda, Customer customer)
         {
-            double valueOfPayment = TotalCoinValue(payment);
-            double amountOfChange = DetermineChange(valueOfPayment, chosenSoda.Price);
+            decimal valueOfPayment = TotalCoinValue(payment);
+            decimal amountOfChange = DetermineChange(valueOfPayment, chosenSoda.Price);
             if (valueOfPayment < chosenSoda.Price)
             {
                 customer.AddCoinsIntoWallet(payment);
@@ -146,7 +146,7 @@ namespace SodaMachine
         //Attempts to gather all the required coins from the sodamachine's register to make change.
         //Returns the list of coins as change to despense.
         //If the change cannot be made, return null.
-        private List<Coin> GatherChange(double changeValue)
+        private List<Coin> GatherChange(decimal changeValue)
         {
             List<Coin> changeToDispense = new List<Coin>();
             while (changeValue > 0)
@@ -197,15 +197,15 @@ namespace SodaMachine
             
         }
         //Takes in the total payment amount and the price of can to return the change amount.
-        private double DetermineChange(double totalPayment, double canPrice)
+        private decimal DetermineChange(decimal totalPayment, decimal canPrice)
         {
-            double changeValue = (totalPayment - canPrice);
+            decimal changeValue = (totalPayment - canPrice);
             return changeValue;
         }
         //Takes in a list of coins to returnt he total value of the coins as a double.
-        private double TotalCoinValue(List<Coin> payment)
+        private decimal TotalCoinValue(List<Coin> payment)
         {
-            double totalValue = 0;
+            decimal totalValue = 0;
             foreach (Coin coin in payment)
             {
                 totalValue += coin.Value;
